+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Cere Stats
 
 ## vNext
-- Updated Cere Blockchain public URLs
+- [BE] Implemented balances monitoring API
+- [BE, FE] Updated Cere Blockchain public URLs
 
 ## v0.16.0
 - [BE, FE] Added DDC metrics